add bools, make numbers floats, add float parsing

+++ src/ast.rs
@@ -3,7 +3,7 @@ use std::rc::Rc;
 #[derive(Debug)]
 pub enum Expression {
     String(String),
-    Number(i32),
+    Number(f64),
     Lookup(String),
     Call(Vec<Expression>),
     Quote(Rc<Vec<Expression>>),
